Implement setup/teardown for fake Comm in __init__.py
